+++ cmake/options.cmake
@@ -74,3 +74,12 @@ option (WITH_FUNC_LINE_LOG "Log with function name, line number prefix" OFF)
 option (WITH_DOC "Build with documentation" ON)
 
 set_property (GLOBAL PROPERTY "PROJECT_EC_FLAGS" -Wall -Werror -Wextra)
+
+if ("${PROJECT_MACHINE}" STREQUAL "zynqmp_a53" OR
+    "${PROJECT_MACHINE}" STREQUAL "zynqmp_a72" OR
+    "${PROJECT_MACHINE}" STREQUAL "zynqmp_a78" OR
+    "${PROJECT_MACHINE}" STREQUAL "zynqmp_r5"  OR
+    "${PROJECT_MACHINE}" STREQUAL "microblaze_generic"  OR
+    "${PROJECT_MACHINE}" STREQUAL "zynq7")
+  add_definitions( -DXLNX_PLATFORM )
+endif()
