+++ lib/system/generic/sys.h
@@ -23,7 +23,11 @@
 #include <stdarg.h>
 #include <string.h>
 
+#ifdef XLNX_PLATFORM
+#include <metal/system/generic/xlnx/sys.h>
+#else
 #include "./@PROJECT_MACHINE@/sys.h"
+#endif
 
 #ifdef __cplusplus
 extern "C" {
