+++ lib/system/generic/xlnx/microblaze_generic/CMakeLists.txt
@@ -1,10 +1,6 @@
-collect (PROJECT_LIB_HEADERS sys.h)
-
 collect (PROJECT_LIB_SOURCES sys.c)
 
 check_include_files(xintc.h HAS_XINTC)
 if (HAS_XINTC)
   add_definitions(-DHAS_XINTC)
 endif(HAS_XINTC)
-
-add_subdirectory(../xlnx_common ${CMAKE_CURRENT_BINARY_DIR}/../xlnx_common)
