--- lib/system/generic/xlnx_common/CMakeLists.txt
@@ -1,7 +1,0 @@
-collect (PROJECT_LIB_HEADERS sys.h)
-
-collect (PROJECT_LIB_SOURCES irq.c)
-
-if ("${PROJECT_MACHINE}" STREQUAL "zynqmp_a53" OR "${PROJECT_MACHINE}" STREQUAL "zynqmp_a72")
-  add_subdirectory(zynqmp_aarch64)
-endif ("${PROJECT_MACHINE}" STREQUAL "zynqmp_a53" OR "${PROJECT_MACHINE}" STREQUAL "zynqmp_a72")
