--- lib/system/generic/xlnx_common/zynqmp_aarch64/CMakeLists.txt
@@ -1,2 +1,0 @@
-collect (PROJECT_LIB_HEADERS sys.h)
-collect (PROJECT_LIB_SOURCES sys.c)
