--- lib/system/generic/zynq7/CMakeLists.txt
@@ -1,5 +1,0 @@
-collect (PROJECT_LIB_HEADERS sys.h)
-
-collect (PROJECT_LIB_SOURCES sys.c)
-
-add_subdirectory(../xlnx_common ${CMAKE_CURRENT_BINARY_DIR}/../xlnx_common)
