--- lib/system/generic/zynqmp_a53/CMakeLists.txt
@@ -1,3 +1,0 @@
-collect (PROJECT_LIB_HEADERS sys.h)
-
-add_subdirectory(../xlnx_common ${CMAKE_CURRENT_BINARY_DIR}/../xlnx_common)
