--- lib/system/generic/zynqmp_a53/sys.h
@@ -1,16 +1,0 @@
-/*
- * Copyright (c) 2015, Xilinx Inc. and Contributors. All rights reserved.
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-/*
- * @file	generic/zynqmp_a53/sys.h
- * @brief	generic zynqmp_a53 system primitives for libmetal.
- */
-
-/*
- * The header file is still required as generic/sys.h expects
- * "./@PROJECT_MACHINE@/sys.h" to still exist.
- */
-#include <metal/system/@PROJECT_SYSTEM@/xlnx_common/zynqmp_aarch64/sys.h>
