--- lib/system/generic/zynqmp_a72/sys.h
@@ -1,17 +1,0 @@
-/*
- * Copyright (c) 2022, Xilinx Inc. and Contributors. All rights reserved.
- * Copyright (c) 2022-2023 Advanced Micro Devices, Inc. All Rights Reserved.
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-/*
- * @file	generic/zynqmp_a72/sys.h
- * @brief	generic zynqmp_a72 system primitives for libmetal.
- */
-
-/*
- * The header file is still required as generic/sys.h expects
- * "./@PROJECT_MACHINE@/sys.h" to still exist.
- */
-#include <metal/system/@PROJECT_SYSTEM@/xlnx_common/zynqmp_aarch64/sys.h>
